+++ install_offline.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+xmake g --network=private
+xrepo import -i ./packages
+xmake -y