+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 All notable changes to this project will be documented in this file.
 `Material-MC` adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.2.0]
+
+### Added
+- Link neighbors with given distance
+- Heisenberg model with three axes anisotropy
+- Heisenberg model with external magnetic field
 
 ## [0.1.0] - 2021.10.25